add missing methods and correct copy paste bugs
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@5956 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/io/streams/asciistreams.hh
@@ -291,8 +291,8 @@ namespace Dune
         readError();
     }
 
-    /** \copydoc Dune::InStreamInterface::readUnsignedInt */
-    void readUnsignedLong ( unsigned int &value )
+    /** \copydoc Dune::InStreamInterface::readUnsignedLong */
+    void readUnsignedLong (unsigned long &value )
     {
       stream_ >> value;
       if( !valid () )
